+++ src/codec/tests/Codec.test.js
@@ -1,8 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { BaseConverter, Converter, ConverterConfig, Ref, Sexp }
-  from '@this/codec';
+import { BaseCodec, Codec, CodecConfig, Ref, Sexp } from '@this/codec';
 import { AskIf } from '@this/typey';
 
 
@@ -32,7 +31,7 @@ describe('encode()', () => {
       ${Symbol('uninterned')}
       ${Symbol.for('interned')}
       `('self-represents $value', ({ value }) => {
-        const conv = new Converter();
+        const conv = new Codec();
         expect(conv.encode(value)).toBe(value);
       });
     });
@@ -40,7 +39,7 @@ describe('encode()', () => {
     test('wraps functions', () => {
       const florp = () => 123;
 
-      const conv = new Converter();
+      const conv = new Codec();
       const data = conv.encode(florp);
 
       expect(data).toBeInstanceOf(Ref);
@@ -50,7 +49,7 @@ describe('encode()', () => {
     test('wraps proxies of regular objects', () => {
       const florp = { a: 123 };
       const proxy = new Proxy(florp, {});
-      const conv  = new Converter();
+      const conv  = new Codec();
       const data  = conv.encode(proxy);
 
       expect(data).toBeInstanceOf(Ref);
@@ -60,7 +59,7 @@ describe('encode()', () => {
     test('wraps proxies of functions', () => {
       const florp = () => 123;
       const proxy = new Proxy(florp, {});
-      const conv  = new Converter();
+      const conv  = new Codec();
       const data  = conv.encode(proxy);
 
       expect(data).toBeInstanceOf(Ref);
@@ -74,7 +73,7 @@ describe('encode()', () => {
 
       const florp = new Florp();
 
-      const conv = new Converter();
+      const conv = new Codec();
       const data = conv.encode(florp);
 
       expect(data).toBeInstanceOf(Ref);
@@ -85,7 +84,7 @@ describe('encode()', () => {
       test('calls `ENCODE()` exactly once', () => {
         let calledCount = 0;
         class Florp {
-          [BaseConverter.ENCODE]() {
+          [BaseCodec.ENCODE]() {
             calledCount++;
             return 123;
           }
@@ -93,7 +92,7 @@ describe('encode()', () => {
 
         const florp = new Florp();
 
-        const conv = new Converter();
+        const conv = new Codec();
         const data = conv.encode(florp);
         expect(calledCount).toBe(1);
         expect(data).toBe(123);
@@ -102,11 +101,11 @@ describe('encode()', () => {
       test('converts the value returned from `ENCODE()`', () => {
         const theData = [1, 2, 3];
         class Florp {
-          [BaseConverter.ENCODE]() { return theData; }
+          [BaseCodec.ENCODE]() { return theData; }
         }
         const florp = new Florp();
 
-        const conv = new Converter();
+        const conv = new Codec();
         const data = conv.encode(florp);
         expect(data).not.toBe(theData);
         expect(data).toBeFrozen();
@@ -116,7 +115,7 @@ describe('encode()', () => {
 
     describe('on arrays', () => {
       test('returns a frozen array', () => {
-        const conv = new Converter();
+        const conv = new Codec();
         const data = conv.encode([1]);
 
         expect(data).toBeArray();
@@ -126,7 +125,7 @@ describe('encode()', () => {
       test('does not return the same array if given a non-frozen one', () => {
         const orig = ['x'];
 
-        const conv = new Converter();
+        const conv = new Codec();
         const data = conv.encode(orig);
 
         expect(data).not.toBe(orig);
@@ -135,7 +134,7 @@ describe('encode()', () => {
       test('returns a strict-equal array', () => {
         const orig = ['x', [1, 2, 3], { foo: 'bar' }];
 
-        const conv = new Converter();
+        const conv = new Codec();
         const data = conv.encode(orig);
 
         expect(data).toStrictEqual(orig);
@@ -144,7 +143,7 @@ describe('encode()', () => {
       test('returns the same array if given a frozen one whose contents all trivially self-represent', () => {
         const orig = Object.freeze(['x', 123, false, undefined, Symbol('foo')]);
 
-        const conv = new Converter();
+        const conv = new Codec();
         const data = conv.encode(orig);
         expect(data).toBe(orig);
       });
@@ -156,7 +155,7 @@ describe('encode()', () => {
           Object.freeze({ z: Object.freeze([false, true]) })
         ]);
 
-        const conv = new Converter();
+        const conv = new Codec();
         const data = conv.encode(orig);
         expect(data).toBe(orig);
       });
@@ -164,7 +163,7 @@ describe('encode()', () => {
       test('returns a different array if any elements need to become frozen', () => {
         const orig = [1, 2, 3, [1, 2, 3], 4, 5, 6];
 
-        const conv = new Converter();
+        const conv = new Codec();
         const data = conv.encode(orig);
         expect(data).not.toBe(orig);
       });
@@ -175,7 +174,7 @@ describe('encode()', () => {
         orig[7]     = 777;
         orig.length = 10;
 
-        const conv = new Converter();
+        const conv = new Codec();
         const got  = conv.encode(orig);
 
         expect(got).toBeArrayOfSize(10);
@@ -187,7 +186,7 @@ describe('encode()', () => {
         const orig = [1, 2, 3];
         orig.florp = ['like', 'yeah'];
 
-        const conv = new Converter();
+        const conv = new Codec();
         const got  = conv.encode(orig);
 
         expect(got).toBeArrayOfSize(3);
@@ -202,7 +201,7 @@ describe('encode()', () => {
         const orig = ['a', 'b'];
         orig[sym] = 'never';
 
-        const conv = new Converter();
+        const conv = new Codec();
         const got  = conv.encode(orig);
 
         expect(got).toBeArrayOfSize(2);
@@ -215,7 +214,7 @@ describe('encode()', () => {
 
     describe('on plain objects', () => {
       test('returns a frozen plain object', () => {
-        const conv = new Converter();
+        const conv = new Codec();
         const data = conv.encode({ x: 10 });
 
         expect(AskIf.plainObject(data)).toBeTrue();
@@ -225,7 +224,7 @@ describe('encode()', () => {
       test('does not return the same object if given a non-frozen one', () => {
         const orig = { beep: 'boop' };
 
-        const conv = new Converter();
+        const conv = new Codec();
         const data = conv.encode(orig);
 
         expect(data).not.toBe(orig);
@@ -234,7 +233,7 @@ describe('encode()', () => {
       test('returns a strict-equal object', () => {
         const orig = { a: 10, b: 20, c: [true, false] };
 
-        const conv = new Converter();
+        const conv = new Codec();
         const data = conv.encode(orig);
         expect(data).toStrictEqual(orig);
       });
@@ -242,7 +241,7 @@ describe('encode()', () => {
       test('returns the same object if given a frozen one whose contents all trivially self-represent', () => {
         const orig = Object.freeze({ a: 10, b: false, c: Symbol('like') });
 
-        const conv = new Converter();
+        const conv = new Codec();
         const data = conv.encode(orig);
         expect(data).toBe(orig);
       });
@@ -254,7 +253,7 @@ describe('encode()', () => {
           three: Object.freeze({ z: Object.freeze([false, true]) })
         });
 
-        const conv = new Converter();
+        const conv = new Codec();
         const data = conv.encode(orig);
         expect(data).toBe(orig);
       });
@@ -266,7 +265,7 @@ describe('encode()', () => {
           z: { seven: 89 }
         });
 
-        const conv = new Converter();
+        const conv = new Codec();
         const data = conv.encode(orig);
 
         expect(data).not.toBe(orig);
@@ -277,7 +276,7 @@ describe('encode()', () => {
         const orig = { a: 10, b: 20 };
         orig[sym] = 'never';
 
-        const conv = new Converter();
+        const conv = new Codec();
         const got  = conv.encode(orig);
 
         expect(got).toBeObject();
@@ -294,7 +293,7 @@ describe('encode()', () => {
         const value1 = Object.freeze(new Sexp('x', null, 1, 2, 3));
         const value2 = new Ref(['blort']);
 
-        const conv = new Converter();
+        const conv = new Codec();
 
         expect(conv.encode(value1)).toBe(value1);
         expect(conv.encode(value2)).toBe(value2);
@@ -310,7 +309,7 @@ describe('encode()', () => {
           both: [value1, value2]
         };
 
-        const conv = new Converter();
+        const conv = new Codec();
         const got  = conv.encode(data);
 
         expect(got.v1).toBe(value1);
@@ -321,7 +320,7 @@ describe('encode()', () => {
 
       test('copies a non-frozen `Sexp` that requires no inner encoding', () => {
         const value = new Sexp('floop', { a: 10, b: 20 }, 1, 2, 3);
-        const conv  = new Converter();
+        const conv  = new Codec();
         const got   = conv.encode(value);
 
         expect(got).not.toBe(value);
@@ -336,7 +335,7 @@ describe('encode()', () => {
       test('handles class `Error` (simple case)', () => {
         const err = new Error('Oy!');
 
-        const conv = new Converter();
+        const conv = new Codec();
         const got  = conv.encode(err);
 
         expect(got).toBeInstanceOf(Sexp);
@@ -358,7 +357,7 @@ describe('encode()', () => {
         err.name        = 'MuffinError';
         err.blueberries = true;
 
-        const conv = new Converter();
+        const conv = new Codec();
         const got  = conv.encode(err);
 
         expect(got).toBeInstanceOf(Sexp);
@@ -389,10 +388,10 @@ describe('encode()', () => {
   });
 
   describe('with default config, except `freeze === false`', () => {
-    const config = new ConverterConfig({ freeze: false });
+    const config = new CodecConfig({ freeze: false });
 
     test('returns a non-frozen array that did not need encoding, as-is', () => {
-      const conv  = new Converter(config);
+      const conv  = new Codec(config);
       const value = [1, 2, 3];
       const got   = conv.encode(value);
 
@@ -401,7 +400,7 @@ describe('encode()', () => {
     });
 
     test('returns a non-frozen object that did not need encoding, as-is', () => {
-      const conv  = new Converter(config);
+      const conv  = new Codec(config);
       const value = { a: 'hello', b: 'goodbye' };
       const got   = conv.encode(value);
 
@@ -410,7 +409,7 @@ describe('encode()', () => {
     });
 
     test('does not freeze an array that needed encoding', () => {
-      const conv  = new Converter(config);
+      const conv  = new Codec(config);
       const value = [1, 2, 3, new Map()];
       const got   = conv.encode(value);
 
@@ -422,7 +421,7 @@ describe('encode()', () => {
     });
 
     test('does not freeze a plain object that needed encoding', () => {
-      const conv  = new Converter(config);
+      const conv  = new Codec(config);
       const value = { boop: new Map() };
       const got   = conv.encode(value);
 
@@ -433,7 +432,7 @@ describe('encode()', () => {
     });
 
     test('copies an array that is frozen', () => {
-      const conv  = new Converter(config);
+      const conv  = new Codec(config);
       const value = Object.freeze([1, 2, 3]);
       const got   = conv.encode(value);
 
@@ -443,7 +442,7 @@ describe('encode()', () => {
     });
 
     test('copies a plain object that is frozen', () => {
-      const conv  = new Converter(config);
+      const conv  = new Codec(config);
       const value = Object.freeze({ florp: 'like' });
       const got   = conv.encode(value);