+++ src/codec/tests/CodecConfig.test.js
@@ -1,21 +1,21 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { BaseConverter, ConverterConfig } from '@this/codec';
+import { BaseCodec, CodecConfig } from '@this/codec';
 
 
 describe('constructor()', () => {
   describe('for config `dataClasses`', () => {
     test('accepts a valid frozen array as-is', () => {
       const classes = Object.freeze([Map]);
-      const got     = new ConverterConfig({ dataClasses: classes });
+      const got     = new CodecConfig({ dataClasses: classes });
 
       expect(got.dataClasses).toBe(classes);
     });
 
     test('makes a frozen copy of a valid non-frozen array', () => {
       const classes = [Map];
-      const got     = new ConverterConfig({ dataClasses: classes });
+      const got     = new CodecConfig({ dataClasses: classes });
 
       expect(got.dataClasses).not.toBe(classes);
       expect(got.dataClasses).toBeFrozen();
@@ -39,19 +39,19 @@ describe('constructor()', () => {
     ${'unhandled'}
     ${'wrap'}
     `('accepts value `$arg`', ({ arg }) => {
-      expect(() => new ConverterConfig({ [name]: arg })).not.toThrow();
+      expect(() => new CodecConfig({ [name]: arg })).not.toThrow();
     });
 
     test('accepts a callable function', () => {
       const func = (val) => `>>${val}<<`;
-      expect(() => new ConverterConfig({ [name]: func })).not.toThrow();
+      expect(() => new CodecConfig({ [name]: func })).not.toThrow();
     });
 
     test('throws given a proper class (non-callable "function")', () => {
       class SomeClass {
         // @emptyBlock
       }
-      expect(() => new ConverterConfig({ [name]: SomeClass })).toThrow();
+      expect(() => new CodecConfig({ [name]: SomeClass })).toThrow();
     });
 
     test.each`
@@ -63,20 +63,20 @@ describe('constructor()', () => {
     ${['wrap']}
     ${'blorp'} // Not one of the allowed string values.
     `('throws given invalid value `$arg`', ({ arg }) => {
-      expect(() => new ConverterConfig({ [name]: arg })).toThrow();
+      expect(() => new CodecConfig({ [name]: arg })).toThrow();
     });
   });
 
   describe('for config `specialCases`', () => {
     test('accepts `null` (which is _not_ the same as the default)', () => {
-      const got = new ConverterConfig({ specialCases: null });
+      const got = new CodecConfig({ specialCases: null });
 
       expect(got.specialCases).toBeNull();
     });
 
-    test('accepts a `BaseConverter` instance', () => {
-      const conv = new BaseConverter();
-      const got  = new ConverterConfig({ specialCases: conv });
+    test('accepts a `BaseCodec` instance', () => {
+      const conv = new BaseCodec();
+      const got  = new CodecConfig({ specialCases: conv });
 
       expect(got.specialCases).toBe(conv);
     });
@@ -88,7 +88,7 @@ describe('constructor()', () => {
     ${'error'}
     ${'omit'}
     `('accepts value `$arg`', ({ arg }) => {
-      expect(() => new ConverterConfig({ symbolKeyAction: arg })).not.toThrow();
+      expect(() => new CodecConfig({ symbolKeyAction: arg })).not.toThrow();
     });
 
     test.each`
@@ -100,7 +100,7 @@ describe('constructor()', () => {
     ${['error']}
     ${'wrap'} // Not one of the allowed string values.
     `('throws given invalid value `$arg`', ({ arg }) => {
-      expect(() => new ConverterConfig({ symbolKeyAction: arg })).toThrow();
+      expect(() => new CodecConfig({ symbolKeyAction: arg })).toThrow();
     });
   });
 });
@@ -112,13 +112,13 @@ describe('constructor()', () => {
 
 describe('makeLoggingInstance()', () => {
   test('produces an instance of this class given no argument', () => {
-    const got = ConverterConfig.makeLoggingInstance();
-    expect(got).toBeInstanceOf(ConverterConfig);
+    const got = CodecConfig.makeLoggingInstance();
+    expect(got).toBeInstanceOf(CodecConfig);
   });
 
   test('produces an instance of this class which reflects the given options', () => {
-    const got = ConverterConfig.makeLoggingInstance({ freeze: false });
-    expect(got).toBeInstanceOf(ConverterConfig);
+    const got = CodecConfig.makeLoggingInstance({ freeze: false });
+    expect(got).toBeInstanceOf(CodecConfig);
     expect(got.freeze).toBeFalse();
   });
 });