+++ src/quant/export/Moment.js
@@ -1,7 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { BaseConverter, Sexp } from '@this/codec';
+import { BaseCodec, Sexp } from '@this/codec';
 import { MustBe } from '@this/typey';
 
 import { Duration } from '#x/Duration';
@@ -223,7 +223,7 @@ export class Moment {
    *
    * @returns {Sexp} Encoded form.
    */
-  [BaseConverter.ENCODE]() {
+  [BaseCodec.ENCODE]() {
     // Note: This string is included for the convenience of humans who happen to
     // be looking at logs (etc.), but is not actually used when reconstructing
     // an instance.